+++ src/libstore/globals.hh
@@ -85,11 +85,6 @@ public:
     std::vector<Path> nixUserConfFiles;
 
     /**
-     * The directory where the man pages are stored.
-     */
-    Path nixManDir;
-
-    /**
      * File name of the socket the daemon listens to.
      */
     Path nixDaemonSocketFile;
